organize declarations
